+++ crates/swc/tests/fixture/issue-2011/no-minify/output/index.js
@@ -1,22 +1,3 @@
-function _classCallCheck(instance, Constructor) {
-    if (!(instance instanceof Constructor)) {
-        throw new TypeError("Cannot call a class as a function");
-    }
-}
-function _defineProperties(target, props) {
-    for(var i = 0; i < props.length; i++){
-        var descriptor = props[i];
-        descriptor.enumerable = descriptor.enumerable || false;
-        descriptor.configurable = true;
-        if ("value" in descriptor) descriptor.writable = true;
-        Object.defineProperty(target, descriptor.key, descriptor);
-    }
-}
-function _createClass(Constructor, protoProps, staticProps) {
-    if (protoProps) _defineProperties(Constructor.prototype, protoProps);
-    if (staticProps) _defineProperties(Constructor, staticProps);
-    return Constructor;
-}
 function _defineProperty(obj, key, value) {
     if (key in obj) {
         Object.defineProperty(obj, key, {
@@ -32,22 +13,15 @@ function _defineProperty(obj, key, value) {
 }
 var ClassA = function ClassA() {
     "use strict";
-    _classCallCheck(this, ClassA);
 };
 module.exports = (function() {
     var ClassB = /*#__PURE__*/ function() {
         "use strict";
-        function ClassB() {
-            _classCallCheck(this, ClassB);
-        }
-        _createClass(ClassB, [
-            {
-                key: "it",
-                value: function it() {
-                    this.bb = new ClassB.MyA();
-                }
-            }
-        ]);
+        function ClassB() {}
+        var _proto = ClassB.prototype;
+        _proto.it = function it() {
+            this.bb = new ClassB.MyA();
+        };
         return ClassB;
     }();
     _defineProperty(ClassB, "MyA", ClassA);