+++ crates/swc_ecma_transforms_compat/tests/es2022_private_in_object.rs
@@ -75,7 +75,10 @@ fn fixture(input: PathBuf) {
                     }
 
                     "transform-classes" => {
-                        pass = Box::new(chain!(pass, classes(Some(t.comments.clone()))));
+                        pass = Box::new(chain!(
+                            pass,
+                            classes(Some(t.comments.clone()), Default::default())
+                        ));
                     }
 
                     _ => {